fix(columbia_merge): refactor common functions to be used in columbia merger and importer, harvard importer
