+++ cl/corpus_importer/utils.py
@@ -4,6 +4,7 @@ from datetime import date
 from difflib import SequenceMatcher
 from typing import Any, Iterator, Optional, Set
 
+from asgiref.sync import async_to_sync
 from bs4 import BeautifulSoup
 from django.db.models import QuerySet
 from django.db.utils import IntegrityError
@@ -15,7 +16,11 @@ from juriscraper.lib.string_utils import harmonize, titlecase
 from cl.citations.utils import map_reporter_db_cite_type
 from cl.lib.command_utils import logger
 from cl.lib.string_diff import get_cosine_similarity
-from cl.people_db.lookup_utils import find_all_judges
+from cl.people_db.lookup_utils import (
+    find_all_judges,
+    lookup_judges_by_last_name_list,
+)
+from cl.people_db.models import Person
 from cl.search.models import Citation, Docket, Opinion, OpinionCluster
 
 
@@ -641,6 +646,30 @@ def add_citations_to_cluster(cites: list[str], cluster_id: int) -> None:
             logger.warning(
                 f"Reporter mismatch for cluster: {cluster_id} on cite: {cite}"
             )
+
+
+def update_cluster_panel(
+    cluster: OpinionCluster,
+    panel_list: list[str],
+    panel_date: Optional[date] = None,
+) -> None:
+    """Update cluster's panel
+
+    This is done independently since it is a m2m relationship, we collect the
+    corrected names, find the Person ids and then add them to the relation
+
+    :param cluster: the cluster object
+    :param panel_list: list with people names
+    :param panel_date: date used to find people
+    :return: None
+    """
+
+    panel_list = [titlecase(p) for p in panel_list]
+    panel = async_to_sync(lookup_judges_by_last_name_list)(
+        panel_list, cluster.docket.court.id, panel_date, True
+    )
+    if panel:
+        cluster.panel.add(*Person.objects.filter(id__in=[p.id for p in panel]))
 
 
 def get_opinion_text(cluster: OpinionCluster) -> str:
